cli example
